Move ssh key generation into install.sh
This code snippet was required to setup the ssh key before the repo
could be cloned. When issue #1 was closed the requirement to have the
ssh key generated before the install.sh script runs is no longer
neccessary.

In this case it makes sense to generate the ssh key at the end of the
install.sh script.

Closes: #2