Move LLVM & clang installation to top of install.sh
This is the only part of the script that requires root privileges.
Moving up in preparation for the removal of hardcoded references to
/home/ubuntu in issue #1.

Hardcoding ofthe home directory is required be cause the script runs
with sudo which causes all references to the home directory with '~' to
expand to the root user's home directory.