Split install.sh into two separe files
The installation of LLVM and clang remain on this file which has been
renamed to install-llvm-clang.sh.

Commands that configure git and setup Vim have been removed and placed
in a configure-git-vim.sh.

All calls to the regular user's home directory habe been replaced with
'~' and '$HOME.

A line has been added at the end of install-llvm-clang.sh which executes
install-llvm-clang.sh under a regular user account. The name of the user
account is passed to the first script as a command line argument.

The README has been edited to include the $USER parameter which is used
to call the second script.

Closes: #3
Fixes: #4