Change working directory in README
The working directory needs to be changed to the same location where the
bash scripts are located before the second script is called at the end
of install-llvm-clang.sh.

It can't be called from inside of the first script as the home directory
will refer to the root user's working directory.