Add test branch, TEST.md
Currently all code is being committed so that it can be cloned on the
test Raspberry Pi. That means that even code which doesn't work is being
committed and pushed to Github. It's not ideal but it has been the easiest, fastest and
most reliable way to get the bash scripts to run on the test target.

This will be a long living branch that will allow tests to run in a
similar fashion to the final script by copying from the TEST.md readme
file and cloning from a test branch.

Then we can run test runs and only commit code that actually works to
the main branch.