Move system git configs to install-llvm-clang.sh
04fd32e set git configurations that were in configure-git-vim.sh
using the --global flag. This meant that all git configurations
would only be available to the currently logged in user. This was a
workaround because the script failed because system wide configs are
written to /etc/gitconfig and would result it a 'Permission denied'
error.

A better solution is to move system-wide configs to the install script
which runs with elevated permissions and keep the user configs in the
second script.